+++ src/matrix/matrix3.c
@@ -24,21 +24,21 @@ t_matrix3	m3subm(t_matrix4 m, int row, int col)
 	return (res);
 }
 
-double	m3minor(t_matrix3 m, int row, int col)
+float	m3minor(t_matrix3 m, int row, int col)
 {
 	return (m2det(m2subm(m, row, col)));
 }
 
-double	m3cofactor(t_matrix3 m, int row, int col)
+float	m3cofactor(t_matrix3 m, int row, int col)
 {
 	if ((row + col) % 2)
 		return (-m3minor(m, row, col));
 	return (m3minor(m, row, col));
 }
 
-double	m3det(t_matrix3 m)
+float	m3det(t_matrix3 m)
 {
-	double	res;
+	float	res;
 	int		i;
 
 	res = 0;
