+++ src/matrix/matrix4.c
@@ -42,21 +42,21 @@ t_matrix4	m4transpose(t_matrix4 m)
 	return (res);
 }
 
-double	m4minor(t_matrix4 m, int row, int col)
+float	m4minor(t_matrix4 m, int row, int col)
 {
 	return (m3det(m3subm(m, row, col)));
 }
 
-double	m4cofactor(t_matrix4 m, int row, int col)
+float	m4cofactor(t_matrix4 m, int row, int col)
 {
 	if ((row + col) % 2)
 		return (-m4minor(m, row, col));
 	return (m4minor(m, row, col));
 }
 
-double	m4det(t_matrix4 m)
+float	m4det(t_matrix4 m)
 {
-	double	res;
+	float	res;
 	int		i;
 
 	res = 0;
